+++ src/engine/Graphics/Context/ExcaliburGraphicsContextWebGL.ts
@@ -31,6 +31,9 @@ import { CircleRenderer } from './circle-renderer/circle-renderer';
 import { Pool } from '../../Util/Pool';
 import { DrawCall } from './draw-call';
 import { AffineMatrix } from '../../Math/affine-matrix';
+import { Material, MaterialOptions } from './material';
+import { MaterialRenderer } from './material-renderer/material-renderer';
+import { Shader, ShaderOptions } from './shader';
 
 export const pixelSnapEpsilon = 0.0001;
 
@@ -220,6 +223,7 @@ export class ExcaliburGraphicsContextWebGL implements ExcaliburGraphicsContext {
 
     // Setup builtin renderers
     this.register(new ImageRenderer());
+    this.register(new MaterialRenderer());
     this.register(new RectangleRenderer());
     this.register(new CircleRenderer());
     this.register(new PointRenderer());
@@ -295,6 +299,7 @@ export class ExcaliburGraphicsContextWebGL implements ExcaliburGraphicsContext {
         drawCall.state.z = this._state.current.z;
         drawCall.state.opacity = this._state.current.opacity;
         drawCall.state.tint = this._state.current.tint;
+        drawCall.state.material = this._state.current.material;
         drawCall.args = args;
         this._drawCalls.push(drawCall);
       } else {
@@ -372,7 +377,12 @@ export class ExcaliburGraphicsContextWebGL implements ExcaliburGraphicsContext {
       }
       return;
     }
-    this.draw<ImageRenderer>('ex.image', image, sx, sy, swidth, sheight, dx, dy, dwidth, dheight);
+
+    if (this._state.current.material) {
+      this.draw<MaterialRenderer>('ex.material', image, sx, sy, swidth, sheight, dx, dy, dwidth, dheight);
+    } else {
+      this.draw<ImageRenderer>('ex.image', image, sx, sy, swidth, sheight, dx, dy, dwidth, dheight);
+    }
   }
 
   public drawLine(start: Vector, end: Vector, color: Color, thickness = 1) {
@@ -461,6 +471,37 @@ export class ExcaliburGraphicsContextWebGL implements ExcaliburGraphicsContext {
         postprocessor.onUpdate(delta);
       }
     }
+  }
+
+  public set material(material: Material) {
+    this._state.current.material = material;
+  }
+
+  public get material(): Material | null {
+    return this._state.current.material;
+  }
+
+  /**
+   * Creates and initializes the material which compiles the internal shader
+   * @param options
+   * @returns Material
+   */
+  public createMaterial(options: MaterialOptions): Material {
+    const material = new Material(options);
+    material.initialize(this.__gl, this);
+    return material;
+  }
+
+  public createShader(options: Omit<ShaderOptions, 'gl'>): Shader {
+    const gl = this.__gl;
+    const { vertexSource, fragmentSource } = options;
+    const shader = new Shader({
+      gl,
+      vertexSource,
+      fragmentSource
+    });
+    shader.compile();
+    return shader;
   }
 
   clear() {