+++ src/engine/Graphics/Context/shader.ts
@@ -1,4 +1,4 @@
-import { Vector } from '../..';
+import { Color, Logger, Vector } from '../..';
 import { Matrix } from '../../Math/matrix';
 import { getAttributeComponentSize, getAttributePointerType } from './webgl-util';
 
@@ -70,7 +70,7 @@ export interface VertexAttributeDefinition {
 
 export interface ShaderOptions {
   /**
-   * WebGL2RenderingContext this layout will be attached to, these cannot be reused across contexts.
+   * WebGL2RenderingContext this layout will be attached to, these cannot be reused across webgl contexts.
    */
   gl: WebGL2RenderingContext;
   /**
@@ -85,6 +85,7 @@ export interface ShaderOptions {
 
 export class Shader {
   private static _ACTIVE_SHADER_INSTANCE: Shader = null;
+  private _logger = Logger.getInstance();
   private _gl: WebGL2RenderingContext;
   public program: WebGLProgram;
   public uniforms: { [variableName: string]: UniformDefinition } = {};
@@ -201,6 +202,18 @@ export class Shader {
   }
 
   /**
+   * Set an integer uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformInt(name: string, value: number): boolean {
+    return this.trySetUniform('uniform1i', name, ~~value);
+  }
+
+  /**
    * Set an integer array uniform for the current shader
    *
    * **Important** Must call ex.Shader.use() before setting a uniform!
@@ -210,6 +223,18 @@ export class Shader {
    */
   setUniformIntArray(name: string, value: number[]) {
     this.setUniform('uniform1iv', name, value);
+  }
+
+  /**
+   * Set an integer array uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformIntArray(name: string, value: number[]): boolean {
+    return this.trySetUniform('uniform1iv', name, value);
   }
 
   /**
@@ -225,6 +250,18 @@ export class Shader {
   }
 
   /**
+   * Set a boolean uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformBoolean(name: string, value: boolean): boolean {
+    return this.trySetUniform('uniform1i', name, value ? 1 : 0);
+  }
+
+  /**
    * Set a float uniform for the current shader
    *
    * **Important** Must call ex.Shader.use() before setting a uniform!
@@ -237,6 +274,18 @@ export class Shader {
   }
 
   /**
+   * Set a float uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformFloat(name: string, value: number): boolean {
+    return this.trySetUniform('uniform1f', name, value);
+  }
+
+  /**
    * Set a float array uniform for the current shader
    *
    * **Important** Must call ex.Shader.use() before setting a uniform!
@@ -246,6 +295,17 @@ export class Shader {
    */
   setUniformFloatArray(name: string, value: number[]) {
     this.setUniform('uniform1fv', name, value);
+  }
+  /**
+   * Set a float array uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformFloatArray(name: string, value: number[]): boolean {
+    return this.trySetUniform('uniform1fv', name, value);
   }
 
   /**
@@ -261,6 +321,42 @@ export class Shader {
   }
 
   /**
+   * Set a [[Vector]] uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformFloatVector(name: string, value: Vector): boolean {
+    return this.trySetUniform('uniform2f', name, value.x, value.y);
+  }
+
+  /**
+   * Set a [[Color]] uniform for the current shader
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  setUniformFloatColor(name: string, value: Color) {
+    this.setUniform('uniform4f', name, value.r / 255, value.g / 255, value.b / 255, value.a);
+  }
+
+  /**
+   * Set a [[Color]] uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformFloatColor(name: string, value: Color): boolean {
+    return this.trySetUniform('uniform4f', name, value.r / 255, value.g / 255, value.b / 255, value.a);
+  }
+
+  /**
    * Set an [[Matrix]] uniform for the current shader
    *
    * **Important** Must call ex.Shader.use() before setting a uniform!
@@ -270,6 +366,18 @@ export class Shader {
    */
   setUniformMatrix(name: string, value: Matrix) {
     this.setUniform('uniformMatrix4fv', name, false, value.data);
+  }
+
+  /**
+   * Set an [[Matrix]] uniform for the current shader, WILL NOT THROW on error.
+   *
+   * **Important** Must call ex.Shader.use() before setting a uniform!
+   *
+   * @param name
+   * @param value
+   */
+  trySetUniformMatrix(name: string, value: Matrix): boolean {
+    return this.trySetUniform('uniformMatrix4fv', name, false, value.data);
   }
 
   /**
@@ -294,6 +402,39 @@ export class Shader {
       throw Error(`Uniform ${uniformType}:${name} doesn\'t exist or is not used in the shader source code,`+
       ' unused uniforms are optimized away by most browsers');
     }
+  }
+
+  /**
+   * Set any available uniform type in webgl. Will try to set the uniform, will return false if the uniform didn't exist,
+   * true if it was set.
+   *
+   * WILL NOT THROW on error
+   *
+   * For example setUniform('uniformMatrix2fv', 'u_my2x2_mat`, ...);
+   *
+   */
+  trySetUniform<TUniformType extends UniformTypeNames>(
+    uniformType: TUniformType,
+    name: string,
+    ...value: UniformParameters<TUniformType>): boolean {
+    if (!this._compiled) {
+      this._logger.warn(`Must compile shader before setting a uniform ${uniformType}:${name}`);
+      return false;
+    }
+    if (!this.isCurrentlyBound()) {
+      this._logger.warn('Currently accessed shader instance is not the current active shader in WebGL,' +
+      ' must call `shader.use()` before setting uniforms');
+      return false;
+    }
+    const gl = this._gl;
+    const location = gl.getUniformLocation(this.program, name);
+    if (location) {
+      const args = [location, ...value];
+      this._gl[uniformType].apply(this._gl, args);
+    } else {
+      return false;
+    }
+    return true;
   }
 
   private _createProgram(gl: WebGLRenderingContext, vertexShader: WebGLShader, fragmentShader: WebGLShader): WebGLProgram {